Listen on IPv6 as well as IPv4 by default
This is the default behaviour since 1.19.0 in the official docker image for nginx.

The first entrypoint script for docker-nginx (10-listen-on-ipv6-by-default.sh) sets this up.
It also checks that the default file has not been replaced by the user
by comparing it to nginx package by checksum, which doesn't really work for this container
because it provides its own different/separate default config.

Check for IPv6 availability was preserved because
default config doesn't work on IPv6-only infrastructure,
but listening on IPv6 without a check would break compatibility with IPv4-only systems.

See:
* https://github.com/nginxinc/docker-nginx/blob/14e73942f37a7f4c1eb886a28b1e183555b4eacd/entrypoint/10-listen-on-ipv6-by-default.sh
* https://github.com/nginxinc/docker-nginx/issues/394